refactor: Extract Link.getUri() from UnmarkedLinkCheck
