+++ src/main/java/net/boyechko/pdf/autoa11y/document/Link.java
@@ -1,0 +1,42 @@
+/*
+ * PDF-Auto-A11y - Automated PDF Accessibility Remediation
+ * Copyright (C) 2026 Richard Boyechko
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published
+ * by the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ *
+ * You should have received a copy of the GNU Affero General Public License
+ * along with this program.  If not, see <https://www.gnu.org/licenses/>.
+ */
+package net.boyechko.pdf.autoa11y.document;
+
+import com.itextpdf.kernel.pdf.PdfDictionary;
+import com.itextpdf.kernel.pdf.PdfName;
+
+/** Helpers for reading Link annotation data. */
+public final class Link {
+
+    private Link() {}
+
+    /** Returns the URI of a Link annotation's URI action, or null if none. */
+    public static String getUri(PdfDictionary annotDict) {
+        PdfDictionary action = annotDict.getAsDictionary(PdfName.A);
+        if (action != null) {
+            PdfName actionType = action.getAsName(PdfName.S);
+            if (PdfName.URI.equals(actionType)) {
+                var uriObj = action.get(PdfName.URI);
+                if (uriObj != null) {
+                    return uriObj.toString();
+                }
+            }
+        }
+        return null;
+    }
+}
